+++ specialist/dependencies.py
@@ -5,7 +5,7 @@ from fastapi import Depends
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from database.db_helper import db_helper
-from .crud import get_specialist
+from .service import get_specialist
